% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemes.R
\name{aa_recommended}
\alias{aa_recommended}
\title{Recommended schemes by vision type}
\usage{
aa_recommended()
}
\value{
A named character vector with elements \code{normal}, \code{red_green_cvd} and
\code{tritan_cvd}, giving the scheme id recommended for each.
}
\description{
Recommended schemes by vision type
}
\examples{
aa_recommended()
}
